Refactor config maps (#470)
* secrets wip

* secrets wip

* secrets wip

* secrets wip - still a mess

* complete secrets refactor

* clean up code

* basic config maps refactor

* add styling

* generalize key/value fields

* add a config map preset reader

* tiny improvements

* bump image version

* implement patching in resource form

* remove unused onSubmit

* integrate forms into "create" components

* Update core/src/i18n/en.yaml
