+++ core-ui/src/components/Predefined/Create/Secrets/SecretsCreate.scss
