+++ core-ui/src/components/Predefined/Create/Secrets/helpers.js
@@ -1,27 +1,7 @@
-import * as components from '../';
+import { getResourceDefs } from 'shared/helpers/getResourceDefs';
 
-export function getSecretDefs(t, context) {
-  return Object.values(components)
-    .filter(component => component.secrets)
-    .map(component => component.secrets(t, context))
-    .flat();
-}
-
-export function readFromFile() {
-  return new Promise(resolve => {
-    const input = document.createElement('input');
-    input.type = 'file';
-    input.addEventListener('change', e => {
-      const file = e.target.files[0];
-      if (!file) resolve(null);
-      const reader = new FileReader();
-      reader.readAsText(file);
-      reader.onload = e =>
-        resolve({ name: file.name, content: e.target.result });
-    });
-    input.click();
-  });
-}
+export const getSecretDefs = (t, context) =>
+  getResourceDefs('secrets', t, context);
 
 export const mapObjectValues = (fn, obj) =>
   Object.fromEntries(
