+++ core-ui/src/components/Predefined/Create/index.js
@@ -2,7 +2,7 @@ export * from './Namespaces/Namespaces.create.js';
 export * from './AddonsConfigurations/AddonsConfigurations.create.js';
 export * from './RoleBindings/RoleBindings.create.js';
 export * from './OAuthClients/OAuthClients.create.js';
-export * from './ConfigMaps.create.js';
+export * from './ConfigMaps/ConfigMaps.create.js';
 export * from './Deployments/Deployments.create.js';
 export * from './ServiceBindings/ServiceBindings.create';
 export * from './ServiceInstances/ServiceInstances.create';
