+++ core-ui/src/components/Predefined/Details/ConfigMap/ConfigMap.details.js
@@ -1,9 +1,11 @@
 import React from 'react';
-import { ControlledBy, ModalWithForm } from 'react-shared';
 import { Button } from 'fundamental-react';
-import { EditConfigMapForm } from './EditConfigMapForm';
-import { ReadonlyEditorPanel } from '../../../../shared/components/ReadonlyEditorPanel';
 import { useTranslation } from 'react-i18next';
+
+import { ControlledBy, ModalWithForm } from 'react-shared';
+
+import { ConfigMapsCreate } from '../../Create/ConfigMaps/ConfigMaps.create';
+import { ReadonlyEditorPanel } from '../../../../shared/components/ReadonlyEditorPanel';
 
 export const ConfigMapsDetails = ({ DefaultRenderer, ...otherParams }) => {
   const { t, i18n } = useTranslation();
@@ -27,7 +29,7 @@ export const ConfigMapsDetails = ({ DefaultRenderer, ...otherParams }) => {
         confirmText={t('common.buttons.edit')}
         className="modal-size--l"
         renderForm={props => (
-          <EditConfigMapForm
+          <ConfigMapsCreate
             configMap={configMap}
             resourceUrl={otherParams.resourceUrl}
             readonlyName={true}
