--- core-ui/src/components/Predefined/Details/ConfigMap/EditConfigMapForm.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import { useUpdate } from 'react-shared';
-import { createPatch } from 'rfc6902';
-import { ConfigMapForm } from '../../../../shared/components/ConfigMap/ConfigMapForm';
-import { useTranslation } from 'react-i18next';
-
-export function EditConfigMapForm(props) {
-  const { t } = useTranslation();
-
-  const patchRequest = useUpdate();
-  const { configMap, onCompleted, onError, resourceUrl, ...formProps } = props;
-
-  const onSubmit = async updatedConfigMap => {
-    const mergedConfigMap = {
-      ...configMap,
-      ...updatedConfigMap,
-      metadata: { ...configMap.metadata, ...updatedConfigMap.metadata },
-    };
-    try {
-      await patchRequest(resourceUrl, createPatch(configMap, mergedConfigMap));
-      onCompleted(t('config-maps.messages.updated'));
-    } catch (e) {
-      console.warn(e);
-      onError(
-        t('config-maps.messages.updated'),
-        `${t('commons.tooltip.error')} ${e.message}`,
-      );
-    }
-  };
-
-  return (
-    <ConfigMapForm {...formProps} configMap={configMap} onSubmit={onSubmit} />
-  );
-}
