+++ core-ui/src/shared/ResourceForm/components/FormComponents.js
@@ -9,6 +9,7 @@ import {
   ComboboxInput as FdComboboxInput,
   FormInput,
   FormLabel,
+  FormTextarea,
   Button,
   Icon,
   MessageStrip,
@@ -16,9 +17,14 @@ import {
 import { Select as WrappedSelect } from 'shared/components/Select/Select';
 import { Tooltip, K8sNameInput } from 'react-shared';
 import classnames from 'classnames';
-import './FormComponents.scss';
 import { useTranslation } from 'react-i18next';
+
+import { base64Decode, base64Encode, readFromFile } from 'shared/helpers';
+
 import { ResourceFormWrapper } from '../ResourceForm';
+
+import './FormComponents.scss';
+import * as Inputs from './Inputs';
 
 export function CollapsibleSection({
   disabled = false,
@@ -351,14 +357,59 @@ export function TextArrayInput({
 }
 
 export function KeyValueField({
+  actions = [],
+  encodable = false,
   defaultOpen,
   isAdvanced,
+  input = Inputs.Text,
   keyProps = {
     pattern: '([A-Za-z0-9][-A-Za-z0-9_./]*)?[A-Za-z0-9]',
   },
+  readableFromFile = false,
+  lockedKeys = [],
   ...props
 }) {
   const { t } = useTranslation();
+
+  const [valuesEncoded, setValuesEncoded] = useState(false);
+  const [decodeErrors, setDecodeErrors] = useState({});
+
+  const toggleEncoding = () => {
+    setDecodeErrors({});
+    setValuesEncoded(!valuesEncoded);
+  };
+
+  const dataValue = value => {
+    if (!encodable || valuesEncoded) {
+      return value?.val || '';
+    } else {
+      try {
+        return base64Decode(value?.val || '');
+      } catch (e) {
+        decodeErrors[value?.key] = e.message;
+        setDecodeErrors({ ...decodeErrors });
+        setValuesEncoded(true);
+        return '';
+      }
+    }
+  };
+
+  if (encodable) {
+    actions = [
+      ...actions,
+      <Button
+        compact
+        option="transparent"
+        glyph={valuesEncoded ? 'show' : 'hide'}
+        onClick={toggleEncoding}
+      >
+        {valuesEncoded
+          ? t('secrets.buttons.decode')
+          : t('secrets.buttons.encode')}
+      </Button>,
+    ];
+  }
+
   return (
     <MultiInput
       defaultOpen={defaultOpen}
@@ -375,6 +426,7 @@ export function KeyValueField({
         ({ value, setValue, ref, onBlur, focus }) => (
           <FormInput
             compact
+            disabled={lockedKeys.includes(value?.key)}
             key="key"
             value={value?.key || ''}
             ref={ref}
@@ -387,20 +439,77 @@ export function KeyValueField({
             placeholder={t('components.key-value-field.enter-key')}
           />
         ),
-        ({ value, setValue, ref, onBlur, focus }) => (
-          <FormInput
-            compact
-            key="value"
-            value={value?.val || ''}
-            ref={ref}
-            onChange={e => setValue({ ...value, val: e.target.value })}
-            onKeyDown={e => focus(e)}
-            onBlur={onBlur}
-            placeholder={t('components.key-value-field.enter-value')}
-          />
-        ),
+        ({ focus, value, setValue, ...props }) =>
+          input({
+            key: 'value',
+            onKeyDown: e => focus(e),
+            value: dataValue(value),
+            placeholder: t('components.key-value-field.enter-value'),
+            setValue: val =>
+              setValue({
+                ...value,
+                val: valuesEncoded || !encodable ? val : base64Encode(val),
+              }),
+            validationState:
+              value?.key && decodeErrors[value.key]
+                ? {
+                    state: 'error',
+                    text: t('secrets.messages.decode-error', {
+                      message: decodeErrors[value.key],
+                    }),
+                  }
+                : undefined,
+            ...props,
+          }),
+        ({ value, setValue }) =>
+          readableFromFile ? (
+            <Tooltip content={t('common.tooltips.read-file')}>
+              <Button
+                compact
+                className="read-from-file"
+                onClick={() =>
+                  readFromFile().then(result =>
+                    setValue({
+                      key: value?.key || result.name,
+                      val: base64Encode(result.content),
+                    }),
+                  )
+                }
+              >
+                {t('components.key-value-form.read-value')}
+              </Button>
+            </Tooltip>
+          ) : (
+            <></>
+          ),
       ]}
+      actions={actions}
       tooltipContent={t('common.tooltips.key-value')}
+      {...props}
+    />
+  );
+}
+
+export function DataField({ title, ...props }) {
+  const { t } = useTranslation();
+
+  return (
+    <KeyValueField
+      fullWidth
+      readableFromFile
+      className="resource-form__data-field"
+      title={title || t('common.labels.data')}
+      input={({ value, setValue, ...props }) => (
+        <FormTextarea
+          compact
+          key="value"
+          value={value}
+          onChange={e => setValue(e.target.value)}
+          placeholder={t('components.key-value-field.enter-value')}
+          className="value-textarea"
+          {...props}
+        />
+      )}
       {...props}
     />
   );
