+++ core-ui/src/shared/ResourceForm/useCreateResource.js
@@ -1,24 +1,32 @@
 import LuigiClient from '@luigi-project/client';
 import { useNotification } from 'react-shared';
 import { useTranslation } from 'react-i18next';
-import { usePost } from 'react-shared';
+import { usePost, useUpdate } from 'react-shared';
+import { createPatch } from 'rfc6902';
 
 export function useCreateResource(
   singularName,
   pluralKind,
   resource,
+  initialResource,
   createUrl,
   afterCreatedFn,
 ) {
   const { t } = useTranslation();
   const notification = useNotification();
   const postRequest = usePost();
+  const patchRequest = useUpdate();
 
   const defaultAfterCreatedFn = () => {
     notification.notifySuccess({
-      content: t('common.create-form.messages.success', {
-        resourceType: singularName,
-      }),
+      content: t(
+        initialResource
+          ? 'common.create-form.messages.patch-success'
+          : 'common.create-form.messages.create-success',
+        {
+          resourceType: singularName,
+        },
+      ),
     });
     LuigiClient.linkManager()
       .fromContext('namespace')
@@ -29,7 +37,20 @@ export function useCreateResource(
 
   return async () => {
     try {
-      await postRequest(createUrl, resource);
+      if (initialResource) {
+        const mergedResource = {
+          ...initialResource,
+          ...resource,
+          metadata: { ...initialResource.metadata, ...resource.metadata },
+        };
+        await patchRequest(
+          createUrl,
+          createPatch(initialResource, mergedResource),
+        );
+      } else {
+        await postRequest(createUrl, resource);
+      }
+
       if (afterCreatedFn) {
         afterCreatedFn(defaultAfterCreatedFn);
       } else {
@@ -38,10 +59,15 @@ export function useCreateResource(
     } catch (e) {
       console.error(e);
       notification.notifyError({
-        content: t('common.create-form.messages.failure', {
-          resourceType: singularName,
-          error: e.message,
-        }),
+        content: t(
+          initialResource
+            ? 'common.create-form.messages.patch-failure'
+            : 'common.create-form.messages.create-failure',
+          {
+            resourceType: singularName,
+            error: e.message,
+          },
+        ),
       });
       return false;
     }
