--- core-ui/src/shared/components/ConfigMap/createConfigMapInput.js
@@ -1,10 +1,0 @@
-export function createConfigMapInput(name, namespace, labels, data) {
-  return {
-    metadata: {
-      name,
-      namespace,
-      labels,
-    },
-    data,
-  };
-}
