+++ core-ui/src/shared/helpers/index.js
@@ -1,2 +1,3 @@
 export * from './safeBase64';
 export * from './getResourceUrl';
+export * from './readFromFile';
